+++ lib/src/di/app_state.dart
@@ -2,18 +2,18 @@ import 'dart:ui';
 
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
-import '../../backend/fs_interface/domain/entity/setting_data.dart';
-import '../../backend/storage/domain/entity/game_config.dart';
-import '../../backend/storage/domain/entity/preset.dart';
-import '../../backend/storage/domain/usecase/card_color.dart';
-import '../../backend/storage/domain/usecase/dark_mode.dart';
-import '../../backend/storage/domain/usecase/enabled_first.dart';
-import '../../backend/storage/domain/usecase/folder_icon.dart';
-import '../../backend/storage/domain/usecase/game_config.dart';
-import '../../backend/storage/domain/usecase/move_on_drag.dart';
-import '../../backend/storage/domain/usecase/paimon_icon.dart';
-import '../../backend/storage/domain/usecase/run_together.dart';
-import '../../backend/storage/domain/usecase/separate_run.dart';
+import '../backend/fs_interface/domain/entity/setting_data.dart';
+import '../backend/storage/domain/entity/game_config.dart';
+import '../backend/storage/domain/entity/preset.dart';
+import '../backend/storage/domain/usecase/card_color.dart';
+import '../backend/storage/domain/usecase/dark_mode.dart';
+import '../backend/storage/domain/usecase/enabled_first.dart';
+import '../backend/storage/domain/usecase/folder_icon.dart';
+import '../backend/storage/domain/usecase/game_config.dart';
+import '../backend/storage/domain/usecase/move_on_drag.dart';
+import '../backend/storage/domain/usecase/paimon_icon.dart';
+import '../backend/storage/domain/usecase/run_together.dart';
+import '../backend/storage/domain/usecase/separate_run.dart';
 import 'storage.dart';
 
 part 'app_state.g.dart';
