+++ lib/src/di/fs_interface.dart
@@ -1,8 +1,8 @@
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
-import '../../backend/fs_interface/data/repo/fs_interface.dart';
-import '../../backend/fs_interface/domain/repo/fs_interface.dart';
-import '../../backend/storage/domain/usecase/ini_editor_arg.dart';
+import '../backend/fs_interface/data/repo/fs_interface.dart';
+import '../backend/fs_interface/domain/repo/fs_interface.dart';
+import '../backend/storage/domain/usecase/ini_editor_arg.dart';
 import 'storage.dart';
 
 part 'fs_interface.g.dart';
