+++ lib/src/di/nahida_store.dart
@@ -3,12 +3,12 @@ import 'dart:io';
 import 'package:flutter/foundation.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
-import '../../backend/akasha/data/akasha.dart';
-import '../../backend/akasha/domain/entity/akasha.dart';
-import '../../backend/akasha/domain/repo/akasha.dart';
-import '../../backend/akasha/domain/usecase/download_url.dart';
-import '../../backend/mod_writer/data/mod_writer.dart';
-import '../../backend/structure/entity/mod_category.dart';
+import '../backend/akasha/data/akasha.dart';
+import '../backend/akasha/domain/entity/akasha.dart';
+import '../backend/akasha/domain/repo/akasha.dart';
+import '../backend/akasha/domain/usecase/download_url.dart';
+import '../backend/mod_writer/data/mod_writer.dart';
+import '../backend/structure/entity/mod_category.dart';
 
 part 'nahida_store.g.dart';
