+++ lib/src/ui/route/category/category.dart
@@ -10,8 +10,8 @@ import 'package:hooks_riverpod/hooks_riverpod.dart';
 import '../../../backend/fs_interface/domain/usecase/file_system.dart';
 import '../../../backend/structure/entity/mod.dart';
 import '../../../backend/structure/entity/mod_category.dart';
-import '../../di/fs_interface.dart';
-import '../../di/fs_watcher.dart';
+import '../../../di/fs_interface.dart';
+import '../../../di/fs_watcher.dart';
 import '../../route_names.dart';
 import '../../widget/category_drop_target.dart';
 import '../../widget/intrinsic_command_bar.dart';
