+++ lib/src/ui/route/category/mod_card.dart
@@ -14,9 +14,9 @@ import '../../../backend/fs_interface/data/helper/path_op_string.dart';
 import '../../../backend/mod_writer/data/mod_writer.dart';
 import '../../../backend/structure/entity/ini.dart';
 import '../../../backend/structure/entity/mod.dart';
-import '../../di/app_state.dart';
-import '../../di/fs_interface.dart';
-import '../../di/mod_card.dart';
+import '../../../di/app_state.dart';
+import '../../../di/fs_interface.dart';
+import '../../../di/mod_card.dart';
 import '../../util/display_infobar.dart';
 import '../../widget/custom_image.dart';
 import 'ini_widget.dart';
