+++ lib/src/ui/route/home_shell.dart
@@ -14,12 +14,12 @@ import 'package:window_manager/window_manager.dart';
 import '../../backend/app_version/domain/github.dart';
 import '../../backend/fs_interface/data/helper/path_op_string.dart';
 import '../../backend/structure/entity/mod_category.dart';
-import '../di/app_state.dart';
-import '../di/app_version.dart';
-import '../di/exe_arg.dart';
-import '../di/fs_interface.dart';
-import '../di/fs_watcher.dart';
-import '../di/storage.dart';
+import '../../di/app_state.dart';
+import '../../di/app_version.dart';
+import '../../di/exe_arg.dart';
+import '../../di/fs_interface.dart';
+import '../../di/fs_watcher.dart';
+import '../../di/storage.dart';
 import '../route_names.dart';
 import '../util/display_infobar.dart';
 import '../util/open_url.dart';
