+++ lib/src/ui/route/loading.dart
@@ -4,7 +4,7 @@ import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:go_router/go_router.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
-import '../di/storage.dart';
+import '../../di/storage.dart';
 import '../route_names.dart';
 import '../widget/appbar.dart';
