+++ lib/src/ui/route/nahida_store/nahida_store.dart
@@ -11,8 +11,8 @@ import 'package:infinite_scroll_pagination/infinite_scroll_pagination.dart';
 import '../../../backend/akasha/domain/entity/akasha.dart';
 import '../../../backend/fs_interface/data/helper/path_op_string.dart';
 import '../../../backend/structure/entity/mod_category.dart';
-import '../../di/fs_watcher.dart';
-import '../../di/nahida_store.dart';
+import '../../../di/fs_watcher.dart';
+import '../../../di/nahida_store.dart';
 import '../../route_names.dart';
 import '../../util/display_infobar.dart';
 import '../../util/tag_parser.dart';
