+++ lib/src/ui/route/setting.dart
@@ -12,9 +12,9 @@ import 'package:hooks_riverpod/hooks_riverpod.dart';
 import '../../backend/fs_interface/domain/entity/setting_data.dart';
 import '../../backend/storage/domain/entity/game_config.dart';
 import '../../backend/storage/domain/usecase/card_color.dart';
-import '../di/app_state.dart';
-import '../di/app_version.dart';
-import '../di/fs_interface.dart';
+import '../../di/app_state.dart';
+import '../../di/app_version.dart';
+import '../../di/fs_interface.dart';
 import '../route_names.dart';
 import '../widget/game_selector.dart';
 import '../widget/setting_element.dart';
